refactor: interface를 이용한 의존성 분리
