+++ app/src/main/java/woowacourse/shopping/data/CartRepositoryImpl.kt
@@ -1,20 +1,21 @@
 package woowacourse.shopping.data
 
 import woowacourse.shopping.model.Product
+import woowacourse.shopping.model.repository.CartRepository
 
 // TODO: Step2 - CartProductDao를 참조하도록 변경
-class CartRepository {
-
+class CartRepositoryImpl : CartRepository {
     private val cartProducts: MutableList<Product> = mutableListOf()
-    fun addCartProduct(product: Product) {
+
+    override fun addCartProduct(product: Product) {
         cartProducts.add(product)
     }
 
-    fun getAllCartProducts(): List<Product> {
+    override fun getAllCartProducts(): List<Product> {
         return cartProducts.toList()
     }
 
-    fun deleteCartProduct(id: Int) {
+    override fun deleteCartProduct(id: Int) {
         cartProducts.removeAt(id)
     }
 }
