+++ app/src/main/java/woowacourse/shopping/model/repository/CartRepository.kt
@@ -1,0 +1,12 @@
+package woowacourse.shopping.model.repository
+
+import woowacourse.shopping.model.Product
+import woowacourse.shopping.ui.injection.RepositoryDI
+
+interface CartRepository : RepositoryDI {
+    fun addCartProduct(product: Product)
+
+    fun getAllCartProducts(): List<Product>
+
+    fun deleteCartProduct(id: Int)
+}
