+++ app/src/main/java/woowacourse/shopping/model/repository/ProductRepository.kt
@@ -1,0 +1,8 @@
+package woowacourse.shopping.model.repository
+
+import woowacourse.shopping.model.Product
+import woowacourse.shopping.ui.injection.RepositoryDI
+
+interface ProductRepository : RepositoryDI {
+    fun getAllProducts(): List<Product>
+}
